backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
pydantic==2.10.3
pydantic-settings==2.7.0
celery[redis]==5.4.0
redis==5.2.1
python-multipart==0.0.18
httpx==0.28.1
python-dotenv==1.0.1
pgvector==0.3.6
websockets==14.1
python-jose==3.3.0
moviepy==1.0.3
python-pptx==1.0.2
openai==1.58.1
aiofiles==24.1.0
edge-tts==6.1.12
Pillow==10.4.0
